// ui/src/lib/api/generated/model/garment.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */
import type { ColorSwatch } from './colorSwatch';
import type { FabricElement } from './fabricElement';
import type { PatternElement } from './patternElement';

export interface Garment {
  /** @nullable */
  garment_id: string | null;
  /** @nullable */
  piece: string | null;
  colors: ColorSwatch[];
  /** @nullable */
  flat: string | null;
  fabrics: FabricElement[];
  patterns: PatternElement[];
}